Create Day18
You are tasked with creating a Person class in C++ that stores details of a person, such as firstName, lastName, and age. The class should:

Have a constructor that takes firstName, lastName, and age as arguments and initializes the member variables.
Include a method getFullName() that returns the full name of the person (a concatenation of firstName and lastName).
Add a method getDetails() that returns the firstName, lastName, and age in a tuple.
Demonstrate the usage of structured bindings to unpack the tuple returned by getDetails().